+++ src/main/java/com/handwoong/rainbowletter/config/security/oauth/CustomOAuthUserService.java
@@ -1,15 +1,13 @@
 package com.handwoong.rainbowletter.config.security.oauth;
 
+import com.handwoong.rainbowletter.config.security.CustomUserDetails;
+import com.handwoong.rainbowletter.domain.member.model.Member;
+import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
 import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
-
-import com.handwoong.rainbowletter.config.security.CustomUserDetails;
-import com.handwoong.rainbowletter.domain.member.Member;
-
-import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
