+++ src/main/java/com/handwoong/rainbowletter/config/security/oauth/OAuthLoginUserParser.java
@@ -1,9 +1,8 @@
 package com.handwoong.rainbowletter.config.security.oauth;
 
-import org.springframework.security.oauth2.core.user.OAuth2User;
-
 import com.handwoong.rainbowletter.config.security.password.PasswordGenerator;
-import com.handwoong.rainbowletter.dto.member.MemberRegisterRequest;
+import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterRequest;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 
 public interface OAuthLoginUserParser {
     MemberRegisterRequest parseDetails(final OAuth2User user, final PasswordGenerator passwordGenerator);
