+++ src/main/java/com/handwoong/rainbowletter/domain/member/dto/MemberResponse.java
@@ -1,6 +1,6 @@
-package com.handwoong.rainbowletter.dto.member;
+package com.handwoong.rainbowletter.domain.member.dto;
 
-import com.handwoong.rainbowletter.domain.member.MemberRole;
+import com.handwoong.rainbowletter.domain.member.model.MemberRole;
 
 public record MemberResponse(
         Long id,
