+++ src/main/java/com/handwoong/rainbowletter/domain/member/model/MemberRole.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.domain.member;
+package com.handwoong.rainbowletter.domain.member.model;
 
 public enum MemberRole {
     ROLE_USER,
