+++ src/main/java/com/handwoong/rainbowletter/domain/member/service/MemberServiceImpl.java
@@ -1,21 +1,21 @@
-package com.handwoong.rainbowletter.service.member;
+package com.handwoong.rainbowletter.domain.member.service;
 
 import com.handwoong.rainbowletter.config.security.GrantType;
 import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
 import com.handwoong.rainbowletter.config.security.TokenResponse;
-import com.handwoong.rainbowletter.domain.member.Member;
-import com.handwoong.rainbowletter.domain.member.MemberStatus;
+import com.handwoong.rainbowletter.domain.member.dto.ChangePasswordRequest;
+import com.handwoong.rainbowletter.domain.member.dto.ChangePhoneNumberRequest;
+import com.handwoong.rainbowletter.domain.member.dto.FindPasswordDto;
+import com.handwoong.rainbowletter.domain.member.dto.MemberLoginRequest;
+import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterRequest;
+import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterResponse;
+import com.handwoong.rainbowletter.domain.member.dto.MemberResponse;
+import com.handwoong.rainbowletter.domain.member.model.Member;
+import com.handwoong.rainbowletter.domain.member.model.MemberStatus;
+import com.handwoong.rainbowletter.domain.member.repository.MemberRepository;
 import com.handwoong.rainbowletter.dto.mail.EmailDto;
-import com.handwoong.rainbowletter.dto.member.ChangePasswordRequest;
-import com.handwoong.rainbowletter.dto.member.ChangePhoneNumberRequest;
-import com.handwoong.rainbowletter.dto.member.FindPasswordDto;
-import com.handwoong.rainbowletter.dto.member.MemberLoginRequest;
-import com.handwoong.rainbowletter.dto.member.MemberRegisterRequest;
-import com.handwoong.rainbowletter.dto.member.MemberRegisterResponse;
-import com.handwoong.rainbowletter.dto.member.MemberResponse;
 import com.handwoong.rainbowletter.exception.ErrorCode;
 import com.handwoong.rainbowletter.exception.RainbowLetterException;
-import com.handwoong.rainbowletter.repository.member.MemberRepository;
 import com.handwoong.rainbowletter.service.mail.event.SendEmail;
 import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 import lombok.RequiredArgsConstructor;
