+++ src/test/java/com/handwoong/rainbowletter/domain/member/controller/MemberControllerTest.java
@@ -1,4 +1,4 @@
-package com.handwoong.rainbowletter.controller.member;
+package com.handwoong.rainbowletter.domain.member.controller;
 
 import static com.handwoong.rainbowletter.config.security.JwtTokenAuthenticationFilter.AUTHORIZATION_HEADER_KEY;
 import static com.handwoong.rainbowletter.config.security.JwtTokenAuthenticationFilter.AUTHORIZATION_HEADER_TYPE;
@@ -13,12 +13,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.handwoong.rainbowletter.config.security.GrantType;
 import com.handwoong.rainbowletter.config.security.JwtTokenProvider;
 import com.handwoong.rainbowletter.config.security.TokenResponse;
-import com.handwoong.rainbowletter.controller.ControllerTestProvider;
-import com.handwoong.rainbowletter.dto.member.ChangePasswordRequest;
-import com.handwoong.rainbowletter.dto.member.ChangePhoneNumberRequest;
-import com.handwoong.rainbowletter.dto.member.FindPasswordDto;
-import com.handwoong.rainbowletter.dto.member.MemberLoginRequest;
-import com.handwoong.rainbowletter.dto.member.MemberRegisterRequest;
+import com.handwoong.rainbowletter.domain.ControllerTestProvider;
+import com.handwoong.rainbowletter.domain.member.dto.ChangePasswordRequest;
+import com.handwoong.rainbowletter.domain.member.dto.ChangePhoneNumberRequest;
+import com.handwoong.rainbowletter.domain.member.dto.FindPasswordDto;
+import com.handwoong.rainbowletter.domain.member.dto.MemberLoginRequest;
+import com.handwoong.rainbowletter.domain.member.dto.MemberRegisterRequest;
 import com.handwoong.rainbowletter.service.mail.template.EmailTemplateType;
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
